constants: Reference on Constants, Units and Uncertainty

This package provides the 2022 version of the CODATA internationally recommended values of the fundamental physical constants for their use within the R language.

References

Eite Tiesinga, Peter J. Mohr, David B. Newell, and Barry N. Taylor (2024). The 2022 CODATA Recommended Values of the Fundamental Physical Constants (Web Version 9.0). Database developed by J. Baker, M. Douma, and S. Kotochigova. Available at https://physics.nist.gov/cuu/Constants/, National Institute of Standards and Technology, Gaithersburg, MD 20899.

See also

Author

Iñaki Ucar