Lookup for Fundamental Physical Constants

A simple wrapper around grep for exploring the CODATA dataset.

Usage

lookup(pattern, cols = c("symbol", "quantity", "type"), ...)

Arguments

pattern

character string containing a regular expression to be matched (see grep).

cols

columns to perform pattern matching (see codata).

...

additional arguments for grep.

See also

Examples

lookup("planck", ignore.case=TRUE)
#>        symbol                                  quantity
#> 138       nah                     molar Planck constant
#> 199         h                           Planck constant
#> 200       hev                  Planck constant in eV/Hz
#> 201      plkl                             Planck length
#> 202      plkm                               Planck mass
#> 203 plkmc2gev      Planck mass energy equivalent in GeV
#> 204    plktmp                        Planck temperature
#> 205      plkt                               Planck time
#> 231      hbar                   reduced Planck constant
#> 232    hbarev           reduced Planck constant in eV s
#> 233   hbcmevf reduced Planck constant times c in MeV fm
#>                          type        value uncertainty     unit
#> 138          Physico-chemical 3.990313e-10     0.0e+00 J/Hz/mol
#> 199 Universal, Adopted values 6.626070e-34     0.0e+00     J/Hz
#> 200                 Universal 4.135668e-15     0.0e+00    eV/Hz
#> 201                 Universal 1.616255e-35     1.8e-40        m
#> 202                 Universal 2.176434e-08     2.4e-13       kg
#> 203                 Universal 1.220890e+19     1.4e+14      GeV
#> 204                 Universal 1.416784e+32     1.6e+27        K
#> 205                 Universal 5.391247e-44     6.0e-49        s
#> 231 Universal, Adopted values 1.054572e-34     0.0e+00      J*s
#> 232                 Universal 6.582120e-16     0.0e+00     eV*s
#> 233   Universal, Non-SI units 1.973270e+02     0.0e+00   MeV*fm